separate users into user & auth mods
